requirements.txt
autodistill==0.1.2
autodistill_grounded_sam==0.1.1
autodistill_yolov8==0.1.0
ipython==8.12.2
opencv_python==4.7.0.72
Pillow==9.5.0
supervision==0.9.0
torch==2.0.1+cu117
ultralytics==8.0.81
